requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==4.2.6
django-crispy-forms==2.1
Faker==20.0.0
mysqlclient==2.2.0
Pillow==10.1.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
